fixed some errors on tooltips
