--- frontend/src/components/CardList/components/Cards/NoteCard/ActiveNoteCard/index.js
@@ -1,1 +1,0 @@
-export { default as ActiveNoteCard } from "./ActiveNoteCard"
