--- frontend/src/components/CardList/components/Cards/NoteCard/ArchivedNoteCard/index.js
@@ -1,1 +1,0 @@
-export { default as ArchivedNoteCard } from "./ArchivedNoteCard"
