--- frontend/src/components/CardList/components/Cards/NoteCard/components/NoteHeader/index.js
@@ -1,1 +1,0 @@
-export { default as NoteHeader } from "./NoteHeader"
