--- frontend/src/components/CardList/components/Cards/NoteCard1/index.js
@@ -1,1 +1,0 @@
-export { default as NoteCard } from "./NoteCard"
